refactor:  unification of the common tests
- Added  and  functions to handle errors for HTTP/1 and HTTP/2 requests.
- Introduced  function to simplify test setup by returning appropriate HTTP modules and request functions based on the type.
